pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gesture-control-app"
version = "1.0.0"
description = "MediaPipe-based gesture control desktop application"
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "mediapipe==0.10.8",
    "opencv-python==4.8.1.78",
    "numpy==1.24.3",
    "pillow>=10.0.1",
    "google-auth>=2.23.4",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "pyautogui>=0.9.54",
    "pynput>=1.7.6",
    "opencv-contrib-python>=4.8.0",
    "boto3>=1.26.0",
    "customtkinter>=5.2.0",
    "requests>=2.31.0",
]
requires-python = ">=3.8.1"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true